Remove "-hwaccel auto"
The "-hwaccel auto" configuration option is somewhat misunderstood in FFMPEG. This is NOT required to achieve hardware encoding with a supported driver/engine.

This flag, in its current position in the cli args list tells FFMPEG to attempt to use hardware acceleration for Decoding. The setting you have further down  specifying the video coded "-c:v hevc_nvenc" is what triggers hardware encoding to be enabled.

The effect of turning off hardware video decoding is, in theory, more gpu resources available for accelerated encoding. My own testing on a GTX 1660 resulted in a 2x encoding speedup when making this change. (1.1x realtime -> 2.15x realtime, on 4k content).

My GTX 2070 Super seems to have much less effect on encoding speed (6.8x to 7.2x, on 1080p), but it is still a slight improvement.

FFMPEG will fallback to cpu based video decoding with the above change, which is fine IMO because decoding is much less cpu intensive then encoding.

Have a test and see if you see similar results.